+++ src/Concerns/ManagesSubscriptions.php
@@ -52,6 +52,21 @@ trait ManagesSubscriptions
     }
 
     /**
+     * Get the ending date of the trial.
+     *
+     * @param  string  $name
+     * @return \Illuminate\Support\Carbon|null
+     */
+    public function trialEndsAt($name = 'default')
+    {
+        if ($this->onGenericTrial()) {
+            return $this->trial_ends_at;
+        }
+
+        return $this->subscription($name)->trial_ends_at;
+    }
+
+    /**
      * Determine if the Stripe model has a given subscription.
      *
      * @param  string  $name
@@ -77,11 +92,7 @@ trait ManagesSubscriptions
      */
     public function subscription($name = 'default')
     {
-        return $this->subscriptions->sortByDesc(function (Subscription $subscription) {
-            return $subscription->created_at->getTimestamp();
-        })->first(function (Subscription $subscription) use ($name) {
-            return $subscription->name === $name;
-        });
+        return $this->subscriptions()->where('name', $name)->first();
     }
 
     /**