+++ tests/Unit/CustomerTest.php
@@ -16,9 +16,10 @@ class CustomerTest extends TestCase
 
         $this->assertFalse($user->onGenericTrial());
 
-        $user->trial_ends_at = Carbon::tomorrow();
+        $user->trial_ends_at = $tomorrow = Carbon::tomorrow();
 
         $this->assertTrue($user->onGenericTrial());
+        $this->assertSame($tomorrow, $user->trialEndsAt());
 
         $user->trial_ends_at = Carbon::today()->subDays(5);